scripts: add_sdk_options: Add possibility to remove the mcu id for github SDK compatibility
Instead of using cmake includes that are mcu id specific, github SDK
uses generic ones and differences are switched by setting MCUX_DEVICE to
the respective mcu id. Furthermore, github SDK uses west to setup the
SDK and all libs and tools from the SDK repository itself are placed to
'core'.
Additionally, this patch adds the necessary option switch for the above
and one for verbose output (so standard output is reduced again).

Signed-off-by: Steffen Hemer <s.hemer@phytec.de>
Signed-off-by: Dominik Haller <d.haller@phytec.de>